Don't try to open projects without version key, light refactoring
- Move CustomPopUpMenu class to custom_menus.py
- Move recent project qsetting editing functions to helpers
- Fix tests
