+++ tests/mock_helpers.py
@@ -51,8 +51,9 @@ def create_toolboxui():
 
 def create_project(toolbox, project_dir):
     """Creates a project for the given ToolboxUI."""
-    with (mock.patch("spinetoolbox.ui_main.ToolboxUI.update_recent_projects"),):
+    with (mock.patch("spinetoolbox.ui_main.update_recent_projects") as mock_urp,):
         toolbox.create_project(project_dir)
+        mock_urp.assert_called()
 
 
 def create_toolboxui_with_project(project_dir):
